Refactoring of the routes folder into more entities
