+++ app/routes/books_routes.py
@@ -2,28 +2,17 @@
 from app.database.models import (
     Book,
     BookUpdate,
-    BookCreate,
-    Person,
-    Rating,
-    RatingCreate,
-    AuthorshipCreate,
-    ReadCreate
+    BookCreate
 )
 from app.services import (
     CreateBookService,
     GetPersonByIdService,
-    CreateAuthorshipService,
-    CreateReadService,
     UpdateBookService,
     DeleteBookService,
     GetBookByIdService,
-    ListBooksService,
-    ListBooksByAuthorService,
-    ListReadersByBookService,
-    CreateRatingService,
-    ListRatingsByBookService
+    ListBooksService
 )
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, HTTPException
 from app.database.gremlin import GraphDB
 from typing import List, Optional
 
@@ -85,49 +74,3 @@ async def list_books_route(skip: int = 0, limit: int = 10, in_stock: Optional[bo
     book_repository = ListBooksService(graph_database)
     response = await book_repository.execute(skip=skip, limit=limit, in_stock=in_stock)
     return response
-
-
-@router.get("/authors/{author_id}", response_model=List[Book])
-async def list_books_by_author_route(author_id: int, skip: int = 0, limit: int = 10) -> List[Book]:
-    graph_database = GraphDB()
-    book_repository = ListBooksByAuthorService(graph_database)
-    response = await book_repository.execute(author_id, skip=skip, limit=limit)
-    return response
-
-
-@router.get("/{book_id}/readers", response_model=List[Person])
-async def list_readers_by_book_route(book_id: str, skip: int = 0, limit: int = 10) -> List[Person]:
-    graph_database = GraphDB()
-    book_repository = ListReadersByBookService(graph_database)
-    response = await book_repository.execute(book_id, skip=skip, limit=limit)
-    return response
-
-
-@router.get("/{book_id}/ratings", response_model=List[Rating])
-async def list_ratings_by_book_route(book_id: str, skip: int = 0, limit: int = 10) -> List[Rating]:
-    graph_database = GraphDB()
-    book_repository = ListRatingsByBookService(graph_database)
-    response = await book_repository.execute(book_id, skip=skip, limit=limit)
-    return response
-
-
-@router.post("/ratings", response_model=Rating)
-async def create_rating_route(rating: RatingCreate) -> Rating:
-    graph_database = GraphDB()
-    book_repository = CreateRatingService(graph_database)
-    response = await book_repository.execute(rating)
-    return response
-
-
-@router.post("/reads", response_model=None)
-async def create_read_route(read: ReadCreate) -> None:
-    graph_database = GraphDB()
-    book_repository = CreateReadService(graph_database)
-    await book_repository.execute(read)
-
-
-@router.post("/authorships", response_model=None)
-async def create_authorship_route(authorship: AuthorshipCreate) -> None:
-    graph_database = GraphDB()
-    book_repository = CreateAuthorshipService(graph_database)
-    await book_repository.execute(authorship)
